+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+### `v0.1.0` 2014-08-14
+
+ * API: remove `prompt` in favor of `promptLicense` and `promptInfo`
+ * Dependencies: use *Inquirer.js* instead of home-made functions.
+ * Licenses: Add Apache-2.0, BSD-2-Clause, BSD-3-Clause, GPL-2.0, GPL-3.0,
+   CDDL-1.0, EPL-1.0, LGPL-2.1, LGPL-3.0 and MPL-2.0.
+
 ### `v0.0.1` 2014-08-13
 
  * Initial release.